Misc Fixes and Small Improvements

+++ Voat/Voat.Business/Utilities/ThumbGenerator.cs
@@ -5,6 +5,7 @@ using System;
 
 using System.Drawing;
 using System.IO;
+using System.Linq;
 using System.Net;
 using System.Threading.Tasks;
 using System.Web;
@@ -55,7 +56,7 @@ namespace Voat.Utilities
         private const int MaxWidth = 70;
 
         // generate a thumbnail while removing transparency and preserving aspect ratio
-        public static async Task<string> GenerateThumbFromImageUrl(string imageUrl, int timeoutInMilliseconds = 3000)
+        public static async Task<string> GenerateThumbFromImageUrl(string imageUrl, int timeoutInMilliseconds = 3000, bool purgeTempFile = true)
         {
             //TODO: Return NULL if file lenght is zero as thumbnail did not generate to local disk
             var randomFileName = GenerateRandomFilename();
@@ -75,8 +76,11 @@ namespace Voat.Utilities
                 if (Settings.UseContentDeliveryNetwork)
                 {
                     await CloudStorageUtility.UploadBlobToStorageAsync(tempPath, "thumbs");
-                    // delete local file after uploading to CDN
-                    File.Delete(tempPath);
+                    if (purgeTempFile)
+                    {
+                        // delete local file after uploading to CDN
+                        File.Delete(tempPath);
+                    }
                 }
                 return Path.GetFileName(tempPath);
             }
@@ -88,7 +92,7 @@ namespace Voat.Utilities
         }
 
         // store uploaded avatar
-        public static async Task<bool> GenerateAvatar(Image inputImage, string userName, string mimetype)
+        public static async Task<bool> GenerateAvatar(Image inputImage, string userName, string mimetype, bool purgeTempFile = true)
         {
             try
             {
@@ -112,9 +116,11 @@ namespace Voat.Utilities
                 {
                     // upload to CDN
                     await CloudStorageUtility.UploadBlobToStorageAsync(tempAvatarLocation, "avatars");
-
-                    // delete local file after uploading to CDN
-                    File.Delete(tempAvatarLocation);
+                    if (purgeTempFile)
+                    {
+                        // delete local file after uploading to CDN
+                        File.Delete(tempAvatarLocation);
+                    }
                 }
                 return true;
             }
@@ -152,69 +158,39 @@ namespace Voat.Utilities
             return rndFileName;
         }
 
-        // generate a thumbnail if submission is a direct link to image or video
-        public static async Task<string> GenerateThumbFromWebpageUrl(string websiteUrl)
+        
+        public static async Task<string> GenerateThumbFromWebpageUrl(string websiteUrl, bool purgeTempFile = true)
         {
             var extension = Path.GetExtension(websiteUrl);
+            var imageExtensions = new string[] { ".jpg", ".png", ".gif", ".jpeg" };
 
-            // this is a direct link to image
-            if (extension != String.Empty)
+            var thumbFileName = (string)null;
+
+            try
             {
-                if (extension == ".jpg" || extension == ".JPG" || extension == ".png" || extension == ".PNG" || extension == ".gif" || extension == ".GIF")
+                if (imageExtensions.Any(x => x.IsEqual(extension)))
                 {
-                    try
-                    {
-                        var thumbFileName = await GenerateThumbFromImageUrl(websiteUrl);
-                        return thumbFileName;
-                    }
-                    catch (Exception)
-                    {
-                        // thumnail generation failed, skip adding thumbnail
-                        return null;
-                    }
+                    // generate a thumbnail if submission is a direct link to image or video
+                    thumbFileName = await GenerateThumbFromImageUrl(websiteUrl, purgeTempFile: purgeTempFile);
                 }
-
-                // try generating a thumbnail by using the Open Graph Protocol
-                try
+                if (String.IsNullOrEmpty(thumbFileName))
                 {
                     var graphUri = new Uri(websiteUrl);
                     var graph = OpenGraph.ParseUrl(graphUri, userAgent: "Voat.co OpenGraph Parser");
 
                     // open graph failed to find og:image element, abort thumbnail generation
-                    if (graph.Image == null)
-                        return null;
-
-                    var thumbFileName = await GenerateThumbFromImageUrl(graph.Image.ToString());
-                    return thumbFileName;
+                    if (graph.Image != null)
+                    {
+                        thumbFileName = await GenerateThumbFromImageUrl(graph.Image.ToString(), purgeTempFile: purgeTempFile);
+                    }
                 }
-                catch (Exception ex)
-                {
-                    EventLogger.Log(ex);
-                    // thumnail generation failed, skip adding thumbnail
-                    return null;
-                }
-            }
-
-            // this is not a direct link to an image, it could be a link to an article or video
-            // try generating a thumbnail by using the Open Graph Protocol
-            try
-            {
-                var graphUri = new Uri(websiteUrl);
-                var graph = OpenGraph.ParseUrl(graphUri, userAgent: "Voat.co OpenGraph Parser");
-
-                // open graph failed to find og:image element, abort thumbnail generation
-                if (graph.Image == null)
-                    return null;
-
-                var thumbFileName = await GenerateThumbFromImageUrl(graph.Image.ToString());
-                return thumbFileName;
             }
             catch (Exception ex)
             {
-                EventLogger.Log(ex);
                 // thumnail generation failed, skip adding thumbnail
-                return null;
+                EventLogger.Log(ex);
             }
+            return thumbFileName;
         }
     }
 }
